+++ pastas/model.py
@@ -38,10 +38,10 @@ from pastas.typing import NoiseModel as NoiseModelType
 from pastas.typing import Solver, StressModel
 from pastas.typing import TimeSeries as TimeSeriesType
 from pastas.typing import TimestampType
-from pastas.ts_processing import (
+from pastas.timeseries_utils import (
     _get_dt,
     _get_time_offset,
-    frequency_is_supported,
+    _frequency_is_supported,
     get_sample,
 )
 from pastas.utils import validate_name
@@ -696,7 +696,7 @@ class Model:
 
         # Set the frequency & warmup
         if freq:
-            self.settings["freq"] = frequency_is_supported(freq)
+            self.settings["freq"] = _frequency_is_supported(freq)
 
         if warmup is not None:
             self.settings["warmup"] = Timedelta(warmup, "D")
@@ -2013,7 +2013,7 @@ class Model:
         series: bool or str, optional
             Export the simulated series or not. If series is "original", the
             original series are exported, if series is "modified",
-            the series are exported after being changed with the timeseries
+            the series are exported after being changed with the time series
             settings. Default is True.
         **kwargs:
             any argument that is passed to :mod:`pastas.io.dump`.
